# Speedup startup and toolbox operations