--- a/lib/galaxy/config.py
+++ b/lib/galaxy/config.py
@@ -23,6 +23,7 @@ from six.moves import configparser
 
 from galaxy.containers import parse_containers_config
 from galaxy.exceptions import ConfigurationError
+from galaxy.queue_worker import send_local_control_task
 from galaxy.util import ExecutionTimer
 from galaxy.util import listify
 from galaxy.util import string_as_bool
@@ -922,6 +923,7 @@ class ConfiguresGalaxyMixin:
         from galaxy.managers.citations import CitationsManager
         from galaxy.tools.deps import containers
         from galaxy.tools.toolbox.lineages.tool_shed import ToolVersionCache
+        import galaxy.tools.search
 
         self.citations_manager = CitationsManager( self )
         self.tool_version_cache = ToolVersionCache(self)
@@ -932,7 +934,9 @@ class ConfiguresGalaxyMixin:
         if self.config.migrated_tools_config not in tool_configs:
             tool_configs.append( self.config.migrated_tools_config )
         self.toolbox = tools.ToolBox( tool_configs, self.config.tool_path, self )
-        self.reindex_tool_search()
+        # Since indexing the toolbox search index can take a while we
+        # send it as a local control task and do not block here.
+        send_local_control_task(self, 'rebuild_toolbox_search_index')
 
         galaxy_root_dir = os.path.abspath(self.config.root)
         file_path = os.path.abspath(getattr(self.config, "file_path"))
@@ -948,16 +952,13 @@ class ConfiguresGalaxyMixin:
             involucro_auto_init=self.config.involucro_auto_init,
         )
         self.container_finder = containers.ContainerFinder(app_info)
+        index_help = getattr(self.config, "index_tool_help", True)
+        self.toolbox_search = galaxy.tools.search.ToolBoxSearch(self.toolbox, index_help)
 
     def reindex_tool_search( self ):
         # Call this when tools are added or removed.
-        import galaxy.tools.search
-        index_help = getattr( self.config, "index_tool_help", True )
-        toolbox = self.toolbox
-        if not hasattr(self, 'toolbox_search'):
-            self.toolbox_search = galaxy.tools.search.ToolBoxSearch( toolbox, index_help )
-        else:
-            self.toolbox_search.update_index(tool_cache=self.tool_cache)
+        self.toolbox_search.build_index(tool_cache=self.tool_cache)
+        self.tool_cache.reset_status()
 
     def _configure_tool_data_tables( self, from_shed_config ):
         from galaxy.tools.data import ToolDataTableManager