--- a/lib/galaxy/datatypes/dataproviders/hierarchy.py
+++ b/lib/galaxy/datatypes/dataproviders/hierarchy.py
@@ -88,7 +88,7 @@ class XMLDataProvider( HierarchalDataProvider ):
         :param  max_depth:  the number of generations of descendents to return
         """
         if not isinstance( max_depth, int ) or max_depth >= 1:
-            for child in element.getchildren():
+            for child in element:
                 child_data = self.element_as_dict( child )
 
                 next_depth = max_depth - 1 if isinstance( max_depth, int ) else None