--- a/lib/galaxy/queue_worker.py
+++ b/lib/galaxy/queue_worker.py
@@ -5,7 +5,6 @@ reloading the toolbox, etc., across multiple processes.
 
 import logging
 import threading
-import time
 
 import galaxy.queues
 from galaxy import util