@bgruening mentioned slow toolbox reloading in gitter a while back, so I had a look at what where major things that could be sped up.

The most effective speedup for reloading the toolbox on my dev instance is rebuilding the search index only for the tools that have changed since the toolbox reloads. The speedup varies with how many tools are installed of course, for me this went from ~2-3 seconds to 0.4 seconds.

In terms of startup times the most effective change is caching the ToolShedRepositories for 1 second in 1b6da123f82e2e603f173aeb1b81ef0124ce6276 and always fetching the ToolVersion objects from the cache  in 4d5dd2d. This should be long enough to be able to load the complete toolbox with just a single emitted db query. This reduced the time to build the full galaxy app object from ~ 7.7 seconds to ~ 4.7 seconds. The more tools you have the better the speedup, I guess.

Data table reloading was also quite slow. It turns out there were hundreds of `os.path.exists()` calls, that even on my SSD were taking quite a while to finish. I addressed that in 12f7dcaf4eec0f20b3400783fb9ce1532401f934

I think this may even speedup other areas in galaxy, just running the unittests went from 140 seconds to 80 seconds locally. I guess everything that accesses a bunch of tools sequentially will benefit from this.