@jmchilton 2017-04-10
A few things:

- > This reduced the time to build the full galaxy app object from ~ 7.7 seconds to ~ 4.7 seconds.

  Wow - that is amazing - very nice!

- The API test are indicating there is something wrong with the data table loading I think.
- I opened a PR against your branch to normalize the log messages with other time tracking I do for various tool stuff - I use a fixed pattern to parse out timing information - there are some scripts in Galaxy to parse this stuff out. I think the pattern makes the log statements in Python look a little nicer as well.