@jmchilton 2017-04-17
Opps sorry about closing that - I meant to cancel my half written comment not close your PR.