@mvdbeek 2017-04-19
> Leaves my startup time at ~8s.

Hmm, that's a bit surprising, because that means the startup actually became slower, since the toolbox indexing doesn't count towards startup time anymore, or did you include the indexing in the timing?

Are you using sqlite locally ? It's possible that the ToolShedRepositoryCache doesn't do much for you if you are not using sqlite.

> Decreases my tool installation index rebuilding down to 33ms from ~700ms.

That was the primary objective, so hooray :)

> @mvdbeek Can we somehow clean the async queue for index rebuilding if Galaxy goes down? It is useless at that point.

Did you actually manage to queue up indexing form a previous restart?
I think we can check that the toolbox had been reloaded at least once before executing the task.
Which is not the same as cleaning the queue. @dannon do you have a suggestion of how to do that?