@martenson 2017-04-19
@mvdbeek I have postgres on a very quick machine with ssd.

I included the index in timing. So ~6.7 without I guess.

And yeah, after one restart I had three queued up indexing tasks.