@mvdbeek 2017-04-19
> And yeah, after one restart I had three queued up indexing tasks.

OK, we should definitely address that before merging. And I have seen an instance on my server where the search index had disappeared (no tools found by typing in the search box) until I triggered a toolbox reload.