@martenson 2017-04-27
@mvdbeek Did I get it right at the call, that the queueing of indexing tasks has been removed from this and will be an independent PR? Do we want to remove WIP from this then?