@mvdbeek 2017-05-01
OK, I am building the toolbox search index during startup as we used to (https://github.com/galaxyproject/galaxy/pull/3909/commits/af156a8f7496dc9a9e70d3992f227c626aec7fa2), because that didn't work on my production instance, which would start up with an empty tool search and only fill the search once the toolbox had been reloaded once. I am keeping track of how many times the toolbox had been reloaded, and reindex the search only if it has been rebuild less often than the toolbox. This way we ignore queued up rebuild requests. I think this is now ready for  a final round of review (ping @martenson).